Moved the instrumentation of fingerprinting apis to a separate file
